+++ application/config/routes.php
@@ -53,6 +53,11 @@ $route['default_controller'] = 'home';
 $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;
 
-
+// Portfolio Routes
 $route["portfolio"] = "Portfolio";
 $route["[pP]ortfolio/(:any)"] = "Portfolio/getSpecificPortfolio/$1";
+
+// Stocks Routes
+$route["stock"] = "Stock";
+$route["stock/getResult"] = "Stock/getSpecificStock";
+