+++ application/views/Stock_All.php
@@ -1,0 +1,10 @@
+<div id="stocks">
+  <form method="POST" action="stock/getResult">
+  	<select name="stockChoice">
+          {stockList}
+          <option value="{Name}">{Name}</option>
+          {/stockList}
+      </select>
+      <input type="submit">
+  </form>
+</div>